- 옷 등록 테스트코드 작성 - 옷 단일 조회 테스트코드 작성 - 옷 전체 조회 테스트코드 작성 - 옷 대분류별 조회 테스트코드 작성

+++ src/main/java/com/wrongweather/moipzy/domain/clothImg/ClothImageRepository.java
@@ -2,5 +2,5 @@ package com.wrongweather.moipzy.domain.clothImg;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClothImageRepository extends JpaRepository<ClothImage, Long> {
+public interface ClothImageRepository extends JpaRepository<ClothImage, Integer> {
 }
